(enhance) Adds LocalService Parameter to Connect-NexusServer
This new parameter adds the ability to connect to
a service running on the local machine.

It adds several private functions that look up if a local
service exists, and then a -LocalService switch to the
Connect-NexusServer function that uses the calculated
values.

This should allow people to more reliably connect locally, without forgetting interesting switches.