(build) Upgrades actions/upload-artifact to v4 and fixes GitVersion Task
v2 of the action was deprecated, and now causes build failures.

For further details, see: https://github.blog/changelog/2024-02-13-deprecation-notice-v1-and-v2-of-the-artifact-actions/

Fixes a GitVersion Issue in the build, where the shallow checkout
of the repository was causing GitVersion to fail.

Finally, adds a quick VSCode task to build the module.